Debugging app was created in 2015, National Level Techfest of KNIT
This is a client for Debugging Game. Teams has to debug codes in given time interval. A cache of questions residues on client machines. Scores are computed loacally at the end of contest responses are submitted to server. Realtime ranking is published by server. This application consits of two modules level one and level 2.  Winner of first round goes to second round.